requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
ase==3.23.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.3.0
cramjam==2.9.0
cycler==0.12.1
datasets==3.1.0
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1731044880986/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dill==0.3.8
diskcache==5.6.3
docker-pycreds==0.4.0
e3nn==0.5.3
evaluate==0.4.3
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1720869315914/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1725214404607/work
fairchem-core==1.2.0
fastparquet==2024.5.0
fastprogress==1.0.3
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.67.1
huggingface-hub==0.26.2
idna==3.10
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1726082825846/work
importlib_resources==6.4.5
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1729866374957/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1731317204262/work
Jinja2==3.1.4
joblib==1.4.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1726610684920/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
kiwisolver==1.4.7
latexcodec==3.0.0
llvmlite==0.43.0
lmdb==1.5.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
monty==2024.10.21
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opt-einsum-fx==0.1.4
opt_einsum==3.4.0
orjson==3.10.10
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
palettable==3.3.3
pandas==2.2.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==11.0.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1726613481435/work
plotly==5.24.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1727341649933/work
propcache==0.2.0
protobuf==5.28.3
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1729847040822/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
PubChemPy==1.0.4
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
pyarrow==18.0.0
pybtex==0.24.0
pyfaidx==0.8.1.3
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pymatgen==2024.10.29
pyparsing==3.2.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytoda==1.1.6
pytz==2024.2
PyYAML==6.0.2
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1728642224099/work
rdkit==2024.3.5
regex==2024.9.11
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
selfies==2.1.2
sentencepiece==0.2.0
sentry-sdk==2.17.0
setproctitle==1.3.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==7.0.5
SmilesPE==0.0.3
smmap==5.0.1
spglib==2.5.0
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
submitit==1.5.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
-e git+ssh://git@github.ibm.com/AI4SD/textcat.git@d13c63a4904414bdec2ccff1a359e47b0375193b#egg=textcat
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.1
toml==0.10.2
torch==2.5.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1724960687225/work
tqdm==4.66.6
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
transformers==4.46.1
triton==3.1.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.2
uncertainties==3.2.2
Unidecode==1.3.8
upfp==0.0.5
urllib3==2.2.3
wandb==0.18.5
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==3.0.6
wrapt==1.17.0
xgboost==2.1.3
xxhash==3.5.0
yarl==1.17.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1731262100163/work
